feat: :ambulance: Tenth Exercise
I just remove the previous code with the objective to add the tenth exercise, this one is a cyclical program that captures a data number each time, and accumulate them. The program will stops when the user types a zero. The program must show: THE SUM OF THE VALUES, THE VALUE OF THE AVERAGE, HOW MANY VALUES WERE ENTERED, GREATER VALUE AND LESS VALUE.